use int as type for options

+++ src/Core/SharedKernels/DataCollection/DataCollection/Views/Questionnaire/RosterTitleQuestionDescription.cs
@@ -5,13 +5,13 @@ namespace WB.Core.SharedKernels.DataCollection.Views.Questionnaire
 {
     public class RosterTitleQuestionDescription
     {
-        public RosterTitleQuestionDescription(Guid questionId, Dictionary<decimal, string> options = null)
+        public RosterTitleQuestionDescription(Guid questionId, Dictionary<int, string> options = null)
         {
             this.QuestionId = questionId;
-            this.Options = options ?? new Dictionary<decimal, string>();
+            this.Options = options ?? new Dictionary<int, string>();
         }
 
         public Guid QuestionId { get; private set; }
-        public Dictionary<decimal, string> Options { get; private set; }
+        public Dictionary<int, string> Options { get; private set; }
     }
 }
