+++ src/Core/SharedKernels/Questionnaire/WB.Core.SharedKernels.Questionnaire/Documents/Option.cs
@@ -7,7 +7,7 @@ namespace Main.Core.Entities.SubEntities
     {
         public Option() { }
 
-        public Option(decimal? value, string title)
+        public Option(int? value, string title)
         {
             this.Value = value;
             this.Title = title;
@@ -16,28 +16,28 @@ namespace Main.Core.Entities.SubEntities
 
         public Option(string value, string title)
         {
-            this.Value = decimal.TryParse(value, NumberStyles.Number, CultureInfo.InvariantCulture, out var val) ? val : null;
+            this.Value = int.TryParse(value, NumberStyles.Number, CultureInfo.InvariantCulture, out var val) ? val : null;
             this.Title = title;
             this.ParentValue = null;
         }
 
-        public Option(decimal? value, string title, decimal? parentValue) 
+        public Option(int? value, string title, int? parentValue) 
             : this(value, title)
         {
             this.ParentValue = parentValue;
         }
         
-        public Option(decimal? value, string title, decimal? parentValue, string? attachmentName)
+        public Option(int? value, string title, int? parentValue, string? attachmentName)
             : this(value, title, parentValue)
         {
             this.AttachmentName = attachmentName;
         }
 
-        public decimal? Value { get; set; }
+        public int? Value { get; set; }
 
         public string Title { get; set; } = String.Empty;
 
-        public decimal? ParentValue { get; set; }
+        public int? ParentValue { get; set; }
         
         public string? AttachmentName { get; set; }
     }
