+++ src/Tests/WB.Tests.Unit.Designer/QuestionnaireTests/CascadingOptions/CascadingOptionsTests.cs
@@ -135,9 +135,9 @@ namespace WB.Tests.Unit.Designer.QuestionnaireTests.CascadingOptions
                 responsibleId: actorId,
                 options: new[]
                 {
-                    new Option(null, String.Empty, (decimal?)null),
-                    new Option(null, String.Empty, (decimal?)null),
-                    new Option(null, String.Empty, (decimal?)null)
+                    new Option(null, String.Empty, (int?)null),
+                    new Option(null, String.Empty, (int?)null),
+                    new Option(null, String.Empty, (int?)null)
                 },
                 linkedToEntityId: null,
                 isFilteredCombobox: false,
